Refactor app modules to new structure
